Much faster index_db load
The current `SELECT COUNT(key) FROM offset_data;` call to obtain the number of sequences in the database is slow for substantial fasta files.

Because the index is read-only, this can be replaced with `SELECT MAX(_ROWID_) FROM offset_data;` for orders of magnitude better performance.

For example, on a fasta with 17M sequences, loading the index with the current `COUNT(key)` call takes about 4 seconds. With `MAX(_ROWID_)` it takes 2.64ms.

On a fasta file with 307M sequences, loading the index with `COUNT(key)` takes 1 minute and 20 seconds. With `MAX(_ROWID_)` it takes 10ms.